Refactor and split code into multiple files
